fts: extract reader/work.rs (MatchWork + cursor-metadata sums)
Phase 1c (1/3). Move the per-query work accounting — MatchWork and the
term/atom posting-byte and planned-range sums that feed op_stats — into
reader/work.rs. MatchWork stays pub and is re-exported from mod.rs
(fts::reader::MatchWork is used by the superfile reader wrapper and the
SQL candidate path); the summing helpers are pub(super). Pure move.
